# ahm92/app/resources/seng365_event_site_api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SENG365 2021: Meetup Site API",
    "version": "1.0.0",
    "description": "This specification documents the API that must be implemented for Assignment 1.\nThe API provides all of the server-side functionality necessary in order to implement the user stories for Assignment 2."
  },
  "servers": [
    {
      "url": "http://seng365-apitest.csse.canterbury.ac.nz:4001/api/v1",
      "description": "reference server"
    }
  ],
...
  "paths": {
...
    "/events/{id}/image": {
      "parameters": [
        {
          "name": "id",
          "description": "ID of the event.",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/EventOverview/properties/eventId"
          }
        }
      ],
      "get": {
        "tags": [
          "events.images"
        ],
        "summary": "Retrieve an event's hero image.",
        "description": "The response MIME type will be either `image/png`, `image/jpeg`, or `image/gif`, depending on the file type of the image being retrieved.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "image": {
                "schema": {
                  "$ref": "#/components/schemas/BinaryImageFile"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "events.images"
        ],
        "summary": "Set an event's hero image.",
        "description": "Only accessible to the organizer of the event. If the event already has a hero image, the current hero image will be replaced with it, and a 200 OK response will be sent. If not, a 201 Created response will be sent.\n\nIf the file is a `.png`, the request will be sent with `Content-Type: image/png`. If the file is a `.jpeg` or `.jpg`, the request will be sent with `Content-Type: image/jpeg`. If the file is a `.gif`, the request will be sent with `Content-Type: image/gif`.\n\nThe request body should contain the raw binary content of the photo, i.e. it should not be in `multipart/form-data` format or contain any file metadata. ",
        "security": [
          {
            "UserToken": []
          }
        ],
        "requestBody": {
          "content": {
            "image/png": {
              "schema": {
                "$ref": "#/components/schemas/BinaryImageFile"
              }
            },
            "image/jpeg": {
              "schema": {
                "$ref": "#/components/schemas/BinaryImageFile"
              }
            },
            "image/gif": {
              "schema": {
                "$ref": "#/components/schemas/BinaryImageFile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/users/{id}": {
      "parameters": [
        {
          "name": "id",
          "description": "ID of the user.",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/FullUser/properties/userId"
          }
        }
      ],
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Retrieve information about a user.",
        "description": "The email field is only included when the currently authenticated user is viewing their **own** details.",
        "security": [
          {},
          {
            "UserToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Change a user's details.",
        "description": "Only accessible by the user themselves. The email must be syntactically valid (e.g. it has to include an `@`, but whether or not it is actually a registered email address will not be checked). The email address must not already be in use. The password must not be an empty string. If the password is being changed, then the current password must also be provided.",
        "security": [
          {
            "UserToken": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "ChangeUserDetailsRequest",
                "properties": {
                  "firstName": {
                    "$ref": "#/components/schemas/User/properties/firstName"
                  },
                  "lastName": {
                    "$ref": "#/components/schemas/User/properties/lastName"
                  },
                  "email": {
                    "$ref": "#/components/schemas/User/properties/email"
                  },
                  "password": {
                    "$ref": "#/components/schemas/FullUser/properties/password"
                  },
                  "currentPassword": {
                    "$ref": "#/components/schemas/FullUser/properties/password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/{id}/image": {
      "parameters": [
        {
          "name": "id",
          "description": "ID of the user.",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/FullUser/properties/userId"
          }
        }
      ],
      "get": {
        "tags": [
          "users.images"
        ],
        "summary": "Retrieve a user's profile image.",
        "description": "The response MIME type will be either `image/png`, `image/jpeg`, or `image/gif`, depending on the file type of the image being retrieved.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "image": {
                "schema": {
                  "$ref": "#/components/schemas/BinaryImageFile"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "users.images"
        ],
        "summary": "Set a user's profile image.",
        "description": "Only accessible to the user themselves. If the user already has a profile photo, the current profile photo will be replaced with it, and a 200 OK response will be sent. If not, a 201 Created response will be sent.\n\nIf the file is a `.png`, the request will be sent with `Content-Type: image/png`. If the file is a `.jpeg` or `.jpg`, the request will be sent with `Content-Type: image/jpeg`. If the file is a `.gif`, the request will be sent with `Content-Type: image/gif`.\n\nThe request body should contain the raw binary content of the photo, i.e. it should not be in `multipart/form-data` format or contain any file metadata. ",
        "security": [
          {
            "UserToken": []
          }
        ],
        "requestBody": {
          "content": {
            "image/png": {
              "schema": {
                "$ref": "#/components/schemas/BinaryImageFile"
              }
            },
            "image/jpeg": {
              "schema": {
                "$ref": "#/components/schemas/BinaryImageFile"
              }
            },
            "image/gif": {
              "schema": {
                "$ref": "#/components/schemas/BinaryImageFile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "users.images"
        ],
        "summary": "Delete a user's profile image.",
        "security": [
          {
            "UserToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "title": "User",
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1,
            "example": "Adam"
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "example": "Anderson"
          },
          "email": {
            "type": "string",
            "minLength": 1,
            "format": "email",
            "example": "aaa11@uclive.ac.nz"
          }
        }
      },
...
      "BinaryImageFile": {
        "type": "string",
        "format": "binary",
        "description": "The raw content of the PNG, JPEG, or GIF image file.",
        "example": "(raw content of PNG, JPEG, or GIF image file)"
      },
...
    },
    "securitySchemes": {
      "UserToken": {
        "type": "apiKey",
        "name": "X-Authorization",
        "in": "header"
      }
    }
  }
}